Explicitly quantify over type variables
Previously, the instantiation code would replace the same universal type
variable with a fresh variable each type it occurred in the type scheme. This
change fixes that by keeping track of which variables are universally
quantified with an explicit generalization step and generating fresh variables
for each of them at the very start of instantiation.